+++ contracts/mocks/nonCompliantToken.sol
@@ -5,31 +5,30 @@ import "../externals/SafeMath.sol";
 
 /// @title NonCompliantToken is a mock ERC20 token that is not compatible with the ERC20 interface.
 contract NonCompliantToken {
-
     using SafeMath for uint256;
 
-    event Approval(address indexed owner, address indexed spender, uint value);
+    event Approval(address indexed owner, address indexed spender, uint256 value);
     event Transfer(address indexed from, address indexed to, uint256 amount);
     /// @dev Total supply of tokens in circulation.
-    uint public totalSupply;
+    uint256 public totalSupply;
 
     /// @dev Balances for each account.
-    mapping(address => uint) public balanceOf;
-    mapping(address => mapping(address => uint)) public allowance;
+    mapping(address => uint256) public balanceOf;
+    mapping(address => mapping(address => uint256)) public allowance;
 
     /// @dev Transfer a token. This throws on insufficient balance.
-    function transfer(address to, uint amount) public {
+    function transfer(address to, uint256 amount) public {
         require(balanceOf[msg.sender] >= amount, "insufficient balance");
         balanceOf[msg.sender] -= amount;
         balanceOf[to] += amount;
         emit Transfer(msg.sender, to, amount);
     }
 
-    function transferFrom(address _from, address _to, uint _value) public {
+    function transferFrom(address _from, address _to, uint256 _value) public {
         if (_to == address(0)) revert();
         if (balanceOf[_from] < _value) revert();
 
-        uint allowed = allowance[_from][msg.sender];
+        uint256 allowed = allowance[_from][msg.sender];
         if (allowed < _value) revert();
 
         balanceOf[_to] = SafeMath.add(balanceOf[_to], _value);
@@ -38,7 +37,7 @@ contract NonCompliantToken {
         emit Transfer(_from, _to, _value);
     }
 
-    function approve(address _spender, uint _value) public {
+    function approve(address _spender, uint256 _value) public {
         //require user to set to zero before resetting to nonzero
         if ((_value != 0) && (allowance[msg.sender][_spender] != 0)) {
             revert();
@@ -49,14 +48,14 @@ contract NonCompliantToken {
     }
 
     /// @dev Credit an address.
-    function credit(address to, uint amount) public returns (bool) {
+    function credit(address to, uint256 amount) public returns (bool) {
         balanceOf[to] += amount;
         totalSupply += amount;
         return true;
     }
 
     /// @dev Debit an address.
-    function debit(address from, uint amount) public {
+    function debit(address from, uint256 amount) public {
         balanceOf[from] -= amount;
         totalSupply -= amount;
     }
